Refactor certificate renewal orchestration and add error handling for evaluation failures
